fix name problems of dir
