--- BraHitYQ/homeworks/Job_2/README.md
@@ -1,3 +1,0 @@
-这是我第二次的作业
-This is my second assignmen
-This is my homework station.I will push my job to there and you can download them.
